Added GjsDebugHooks interface and GjsMultiplexedDebugHooks object.
This interface (and object) serve as a central point for clients to connect
to SpiderMonkey's internal debugging functions without stepping on each other.

Through GjsDebugHooks, clients are able to connect to the general
execution hooks (such as entry and exit of new functions and source files),
enable single-step mode and recieve interrupts and add arbitrary breakpoints
on source files.

Each of these functions takes a callback and returns a GjsDebugConnection.

This object is effectively just a handle for an active connection to the
debugging interface. Once clients are done with a particular debugging hook
they should unref their connection. This will automatically call back into
the interrupt register's implementation and handle the proper tear down
of that hook inside of SpiderMonkey if required.

https://bugzilla.gnome.org/show_bug.cgi?id=721246