Add GjsDebugHooks
This object serves as a central point for clients to connect
to SpiderMonkey's internal debugging functions without stepping on each other.

Through GjsDebugHooks, clients are able to connect to the general
execution hooks (such as entry and exit of new functions and source files),
enable single-step mode and recieve interrupts and add arbitrary breakpoints
on source files.

Each of these functions takes a callback and returns a guint.

Remove the debugging hook with the relevant _remove function.

https://bugzilla.gnome.org/show_bug.cgi?id=721246