Added GjsCoverage object (and thus code coverage support)
This object effectively just takes a connection to the single step
handler and records how many times we hit each line.

To prevent clutter, files considered to be part of the coverage report
must be specified inclusively, anything not part of those paths will not
be included in the coverage report.

The coverage tool is integrated in to gjs-console, you can use it by
specifying -C or --coverage-path to add a file to generate coverage reports
for. If paths are specified, then --coverage-output is mandatory in order
to specify a directory to write coverage reports to. In order to handle
writing reports for URIs that GVfs might understand but lcov would not,
we need to copy the covered files into this directory.

Coverage reports are append-only by default. Just remove the report
if you need to zero the counters.

https://bugzilla.gnome.org/show_bug.cgi?id=721246