Write function line number before unique key in FN entries.
Before we were not writing the line number before the unique key
but instead as part of the unique key. This is technically not
compliant with the lcov specification, which requires th
following:

    FN:<line number>,<function name>

For some reason, genhtml never complained about this and silently
just accepted the function names. Other tools that work with lcov
data, like lcov_cobertura and lcov-result-merge are far less
forgiving.

Fixes BGO #742362