gi: Call init_object_private after creating JSObject for GObject
Previously we called this function in init_func (eg, _init), but
that left open a window for a garbage collection cycle to occurr
because we were getting and setting properties in the GObject
constructor which were calling into JS code.

When the newly constructed but not yet init'd object got traced,
a null pointer was dereferenced.

Fixes #742517