Don't allow exceptions thrown from debugger to propogate.
A common pattern in some test cases is to start a main loop. The
debugger object which keeps track of which lines are executing
for the purposes of coverage counting may throw an exception.
If it does, then the exception will be thrown in both compartments
and the main loop will never exit. This causes tests to hang.

Instead, we should handle the exception gracefully - e.g., log
the error and refuse to count coverage for that file.

Fixes #742797